fix: a new recipe is lost when moving between components
when adding a new recipe, in the recipe component, more exactly in the
recipe service, if we navigate to the shopping list component, and
back to the recipe component, the recipe added is lost.

this happens because we must use the RecipeService as a provider in our
app module to have the same instance for all the components.

before, when we had it in the providers list in the recipe component,
it would get initialized each time we moved in the recipe component and
lost each time we moved out of it